[IR] Some tweaks to the inlining loop; made profiling counter use a long instead of an int.

+++ src/org/jruby/compiler/ir/Counter.java
@@ -3,5 +3,5 @@ package org.jruby.compiler.ir;
 // Just an int-wrapper so that we can modify the counter in-place
 // (unlike an Integer which is immutable)
 public class Counter {
-    public int count = 0;
+    public long count = 0;
 }
